third_party/blink/web_tests/platform/win7/virtual/exotic-color-space/images/feature-policy-oversized-images-expected.txt

CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
CONSOLE ERROR: Feature policy violation: oversized-images is not allowed in this document.
